pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "korean-social-simulator"
version = "0.1.0"
description = "Synthetic Korean social simulation lab using Concordia and Nemotron-Personas-Korea"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Korean Social Simulation Lab"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
]
hf = [
    "datasets>=2.0",
]
polars = [
    "polars>=0.20",
]
rag = [
]
all = [
    "korean-social-simulator[dev,hf,polars]",
]

[project.scripts]
kssim = "korean_social_simulator.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/korean_social_simulator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_hf: tests that require Hugging Face dataset access",
    "live_llm: tests that require live LLM provider access",
    "live_pageindex: tests that require PageIndex MCP access",
    "integration: integration tests",
    "smoke: smoke tests",
    "golden: golden output tests",
    "slow: slow running tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = [
    "tests/",
]